LPS-94301 blogs - classes in module that are not exported should be in internal package
