+++ modules/apps/blogs/blogs-web/src/main/java/com/liferay/blogs/web/internal/constants/BlogsWebConstants.java
@@ -12,17 +12,20 @@
  * details.
  */
 
-package com.liferay.blogs.web.constants;
+package com.liferay.blogs.web.internal.constants;
 
 /**
- * @author Roberto Díaz
+ * @author Adolfo Pérez
  */
-public class BlogsWebKeys {
+public class BlogsWebConstants {
 
-	public static final String BLOG_ENTRIES_DISPLAY_CONTEXT =
-		"BLOG_ENTRIES_DISPLAY_CONTEXT";
+	public static final String BLOGS_ELEMENTS_DEFAULT_EVENT_HANDLER =
+		"BLOGS_ELEMENTS_DEFAULT_EVENT_HANDLER";
 
-	public static final String BLOGS_ITEM_SELECTOR_HELPER =
-		"BLOGS_ITEM_SELECTOR_HELPER";
+	public static final String BLOGS_ENTRY_ASSET_ENTRY =
+		"BLOGS_ENTRY_ASSET_ENTRY";
+
+	public static final String BLOGS_PORTLET_INSTANCE_CONFIGURATION =
+		"BLOGS_PORTLET_INSTANCE_CONFIGURATION";
 
 }
